simplify hit tolerance in the webgl layer renderer
