+++ src/ol/render/canvas/ExecutorGroup.js
@@ -11,6 +11,7 @@ import {
   compose as composeTransform,
   create as createTransform,
 } from '../../transform.js';
+import {getPixelIndexArray} from '../pixelIndexUtil.js';
 import Executor from './Executor.js';
 
 /**
@@ -489,64 +490,6 @@ class ExecutorGroup {
       deferredZIndexContexts[zs[i]].length = 0;
     }
   }
-}
-
-/**
- * This cache is used to store arrays of indexes for calculated pixel circles
- * to increase performance.
- * It is a static property to allow each Replaygroup to access it.
- * @type {Object<number, Array<number>>}
- */
-const circlePixelIndexArrayCache = {};
-
-/**
- * This methods creates an array with indexes of all pixels within a circle,
- * ordered by how close they are to the center.
- * A cache is used to increase performance.
- * @param {number} radius Radius.
- * @return {Array<number>} An array with indexes within a circle.
- */
-export function getPixelIndexArray(radius) {
-  if (circlePixelIndexArrayCache[radius] !== undefined) {
-    return circlePixelIndexArrayCache[radius];
-  }
-
-  const size = radius * 2 + 1;
-  const maxDistanceSq = radius * radius;
-  const distances = new Array(maxDistanceSq + 1);
-  for (let i = 0; i <= radius; ++i) {
-    for (let j = 0; j <= radius; ++j) {
-      const distanceSq = i * i + j * j;
-      if (distanceSq > maxDistanceSq) {
-        break;
-      }
-      let distance = distances[distanceSq];
-      if (!distance) {
-        distance = [];
-        distances[distanceSq] = distance;
-      }
-      distance.push(((radius + i) * size + (radius + j)) * 4 + 3);
-      if (i > 0) {
-        distance.push(((radius - i) * size + (radius + j)) * 4 + 3);
-      }
-      if (j > 0) {
-        distance.push(((radius + i) * size + (radius - j)) * 4 + 3);
-        if (i > 0) {
-          distance.push(((radius - i) * size + (radius - j)) * 4 + 3);
-        }
-      }
-    }
-  }
-
-  const pixelIndex = [];
-  for (let i = 0, ii = distances.length; i < ii; ++i) {
-    if (distances[i]) {
-      pixelIndex.push(...distances[i]);
-    }
-  }
-
-  circlePixelIndexArrayCache[radius] = pixelIndex;
-  return pixelIndex;
 }
 
 export default ExecutorGroup;
